requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
PyMySQL  
Werkzeug